refactor(landing):components names to pascal WIP
